third_party/blink/web_tests/media/media-fragments/TC0028-expected.txt


Title: TC0028
Fragment: 't=3,banana'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
